fix: re organised file structure
