+++ frontend/src/pages/GraphicalSelector/CourseInfo/GraphicalPlannerButton.tsx
