+++ frontend/src/pages/GraphicalSelector/CourseInfo/index.ts
@@ -1,0 +1,3 @@
+import CourseInfo from './CourseInfo';
+
+export default CourseInfo;
