+++ frontend/src/pages/GraphicalSelector/HowToUse/index.ts
@@ -1,0 +1,3 @@
+import HowToUse from './HowToUse';
+
+export default HowToUse;
